+++ spec/support/route_matcher_helpers.rb
@@ -8,6 +8,10 @@ module Spec
           prefix 'api'
           format 'json'
 
+          get 'custom_name', as: :my_custom_route_name do
+            'hello'
+          end
+
           get 'ping' do
             'pong'
           end
@@ -27,6 +31,12 @@ module Spec
           route :any, '*path' do
             'catch-all route'
           end
+        end
+      end
+
+      def self.mounting_api
+        Class.new(Grape::API) do
+          mount Spec::Support::RouteMatcherHelpers.api
         end
       end
     end