+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## September 27 2015
+
+* You can now assign custom helper names to Grape routes
+* Fixed a bug where routes would be listed more than once in the rake task if they are mounted to another API
+* Added the HTTP verb to rake task output
+
 ## July 5 2015
 
 * You can now pass query parameters to helper functions in order to generate your own query string (Issue #1)